_scripts/requirements.txt
openai
pyyaml
ebooklib
hypothesis
pytest
